requirements.txt
dronekit==2.9.0
Flask==0.12
Flask_SocketIO==2.8.4
gevent==1.2.1
pymavlink==2.0.6
cv2==1.0
eventlet==0.20.1
numpy==1.12.0
pyserial==3.2.1
